requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic[email]==2.4.2
PyJWT==2.8.0
bcrypt==4.0.1
opencv-python==4.8.1.78
numpy==1.26.1
face-recognition==1.3.0
Pillow==10.1.0
motor==3.3.1
pymongo==4.6.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
emails==0.6
jinja2==3.1.2
pyotp==2.9.0
